Add addtional_schema field
This adds the additional_schema field to QlogFile defined in the main
schema. The purpose being to explicitly identify the schema used for
additional events (possibly) included in qlog traces.

The field is free-forma and could be anything. We could go XML style
and require a URL. I've started by just proposing that we reference
the document name:

for I-Ds, this makes it easy to manage iterative versions
for RFCs, these have a canonical source already. If an RFC update
is published and an implementation switches to it, we can change
easily.

Closes #283